requirements.txt
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cx-Oracle==8.3.0
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==23.0.0
identity==0.9.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.31.0
msgspec==0.18.6
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.4
